Factor date header out of preview component and into its own file
