+++ client/index.jsx
@@ -4,7 +4,7 @@ import ReactDom from 'react-dom';
 import axios from 'axios';
 import App from './components/app';
 
-const PROJECT_ID = 2;
+const PROJECT_ID = 5;
 
 document.addEventListener('DOMContentLoaded', () => {
   axios.get(`http://localhost:3000/${PROJECT_ID}/updates`).then(response => {
